+++ Wrappers/Python/cil/optimisation/algorithms/FISTA.py
@@ -62,14 +62,6 @@ class FISTA(Algorithm):
         :param g: Convex function with " simple " proximal operator'''
         
         super(FISTA, self).__init__(**kwargs)
-        if kwargs.get('x_init', None) is not None:
-            if initial is None:
-                warnings.warn('The use of the x_init parameter is deprecated and will be removed in following version. Use initial instead',
-                   DeprecationWarning, stacklevel=4)
-                initial = kwargs.get('x_init', None)
-            else:
-                raise ValueError('{} received both initial and the deprecated x_init parameter. It is not clear which one we should use.'\
-                    .format(self.__class__.__name__))
         
         if initial is not None and f is not None:
             self.set_up(initial=initial, f=f, g=g)