+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * 22.0.0
   - Strongly convex functionality in TotalVariation and FGP_TV Functions.
+  - Deprecated: `x_init` removed from the initialisation of the Algorithms and `initial` will be used.
   - DataProcessors use weak-reference to input data
   - Merged CIL-ASTRA code in to CIL repository simplifying test, build and install procedures
     - Modules not moved should be considered deprecated